// src/types/response.ts
export enum RESPONSE_ERROR {
  NOT_FOUND = "Not Found",
  SERVER_ERROR = "Server Error",
  BAD_REQUEST = "Bad Request",
  UNAUTHORIZED = "Unauthorized",
  NOT_AUTHENTICATED = "Not Authenticated",
  FORBIDDEN = "Forbidden",
  VALIDATION_ERROR = "Validation Error",
  TOO_MANY_REQUESTS = "Too Many Requests",
  INVALID_CREDENTIALS = "Invalid Credentials",
  UNSUPPORTED_MEDIA_TYPE = "Unsupported Media Type",
  INCORRECT_LOGIN_DATA = "Incorrect username or password",
}